refactor: move textlink to nav
